Refactor: Extract decodeCallData to utility function
- Create new utility file: frontend/src/utils/callDataDecoder.js
- Remove duplicate decodeCallData function from SignatureConfirmationScreen.jsx
- Remove duplicate decodeCallData function from SignatureConfirmationDialog.jsx
- Both components now import and use the shared utility function
- Improves code maintainability and reduces duplication (~70 lines removed per file)
